better seed

+++ src/CertificateMutator.h
@@ -1,3 +1,4 @@
 #include <string>
+#include <stdint.h>
 
-extern std::string getCert(unsigned int seed_in, const uint8_t *cert_in, size_t certLen);
+extern std::string getCert(uint64_t seed_in, const uint8_t *cert_in, size_t certLen);
