+++ src/deletionmanipulator.h
@@ -26,12 +26,12 @@ limitations under the License.
 */
 class DeletionManipulator : public Manipulator {
 public:
-  DeletionManipulator(shared_ptr<DERObject> obj, unsigned int randomness);
-  void generate(unsigned int randomness, bool random, int index = -1);
+  DeletionManipulator(shared_ptr<DERObject> obj, uint64_t randomness);
+  void generate(uint64_t randomness, bool random, int index = -1);
   size_t get_fixed_manipulations_count();
 
 private:
-  void set_fixed_manipulations(unsigned int randomness);
+  void set_fixed_manipulations(uint64_t randomness);
   vector<vector<size_t>> fixed_manipulations;
   vector<vector<size_t>> get_fixed_manipulations();
 };
