+++ src/generalizedtimemanipulator.h
@@ -26,18 +26,18 @@ limitations under the License.
 class GeneralizedTimeManipulator : public Manipulator {
 public:
   GeneralizedTimeManipulator(shared_ptr<DERObject> obj,
-                             unsigned int randomness);
-  void generate(unsigned int randomness, bool random, int index = -1);
+                             uint64_t randomness);
+  void generate(uint64_t randomness, bool random, int index = -1);
   virtual size_t get_fixed_manipulations_count();
 
   string get_value();
   void set_value(string str);
 
-  string get_random_time(unsigned int randomness);
+  string get_random_time(uint64_t randomness);
 
 private:
   vector<string> fixed_manipulations;
-  void set_fixed_manipulations(unsigned int randomness);
+  void set_fixed_manipulations(uint64_t randomness);
 
   vector<string> get_fixed_manipulations();
