+++ src/insertionmanipulator.h
@@ -26,15 +26,15 @@ limitations under the License.
 */
 class InsertionManipulator : public Manipulator {
 public:
-  InsertionManipulator(shared_ptr<DERObject> obj, unsigned int randomness);
-  void generate(unsigned int randomness, bool random, int index = -1);
+  InsertionManipulator(shared_ptr<DERObject> obj, uint64_t randomness);
+  void generate(uint64_t randomness, bool random, int index = -1);
   size_t get_fixed_manipulations_count();
 
 private:
-  void set_fixed_manipulations(unsigned int randomness);
+  void set_fixed_manipulations(uint64_t randomness);
   vector<tuple<vector<byte>, size_t>> fixed_manipulations;
   vector<tuple<vector<byte>, size_t>>
-  get_fixed_manipulations(unsigned int randomness);
+  get_fixed_manipulations(uint64_t randomness);
 };
 
 #endif
