+++ src/oidmanipulator.h
@@ -24,8 +24,8 @@ limitations under the License.
 */
 class OIDManipulator : public Manipulator {
 public:
-  OIDManipulator(shared_ptr<DERObject> obj, unsigned int randomness);
-  void generate(unsigned int randomness, bool random, int index = -1);
+  OIDManipulator(shared_ptr<DERObject> obj, uint64_t randomness);
+  void generate(uint64_t randomness, bool random, int index = -1);
   size_t get_fixed_manipulations_count();
 
   vector<int> get_value();
@@ -33,7 +33,7 @@ public:
 
 private:
   vector<vector<int>> fixed_manipulations;
-  void set_fixed_manipulations(unsigned int randomness);
+  void set_fixed_manipulations(uint64_t randomness);
 
   vector<vector<int>> get_fixed_manipulations();
