+++ src/printablestringmanipulator.h
@@ -26,8 +26,8 @@ limitations under the License.
 class PrintableStringManipulator : public Manipulator {
 public:
   PrintableStringManipulator(shared_ptr<DERObject> obj,
-                             unsigned int randomness);
-  void generate(unsigned int randomness, bool random, int index = -1);
+                             uint64_t randomness);
+  void generate(uint64_t randomness, bool random, int index = -1);
   virtual size_t get_fixed_manipulations_count();
 
   string get_value();
@@ -35,7 +35,7 @@ public:
 
 private:
   vector<string> fixed_manipulations;
-  void set_fixed_manipulations(unsigned int randomness);
+  void set_fixed_manipulations(uint64_t randomness);
 
   vector<string> get_fixed_manipulations();
